Bug fix and exception handling
Fixed a bug with alt origin signals coupled with alt routes/overlaps
When a NDZ/BLK has no entry signal to the station FP module will halt. Because this situation is actually a layout error, a message will be shown to the user. If the user insists on the design, a fictitious signal must be added, for now.